Moved ChemKED_to_ReSpecTh converter to ChemKED, and more tests for converters
