+++ pyked/converters.py
@@ -14,7 +14,6 @@ import pint
 
 # Local imports
 from .validation import yaml, property_units
-from .chemked import ChemKED, DataPoint
 from .utils import units as unit_registry
 from ._version import __version__
 
@@ -381,7 +380,7 @@ def get_datapoints(root):
     return datapoints
 
 
-def convert_from_ReSpecTh(filename_xml, filename_ck='', file_author='', file_author_orcid=''):
+def ReSpecTh_to_ChemKED(filename_xml, filename_ck='', file_author='', file_author_orcid=''):
     """Convert ReSpecTh XML file to ChemKED YAML file.
 
     Args:
@@ -390,6 +389,8 @@ def convert_from_ReSpecTh(filename_xml, filename_ck='', file_author='', file_aut
         file_author (`str`, optional): Name to override original file author
         file_author_orcid (`str`, optional): ORCID of file author
     """
+    from pyked.chemked import ChemKED
+
     assert os.path.isfile(filename_xml), 'Error: ' + filename_xml + ' file missing'
 
     # get all information from XML file
@@ -456,212 +457,10 @@ def convert_from_ReSpecTh(filename_xml, filename_ck='', file_author='', file_aut
     ChemKED(yaml_file=filename_ck)
 
 
-def convert_to_ReSpecTh(filename_ck, filename_xml=''):
-    """Convert ChemKED file to ReSpecTh XML file.
-
-    This converter uses common information in a ChemKED file to generate a
-    ReSpecTh XML file. Note that some information may be lost, as ChemKED stores
-    some additional attributes.
-
-    Arguments:
-        filename_ck (`str`): Name of ChemKED YAML file to be converted.
-        filename_xml (`str`, optional): Name of output ReSpecTh XML file to be produced.
-    """
-    c = ChemKED(yaml_file=filename_ck)
-
-    root = etree.Element('experiment')
-
-    file_author = etree.SubElement(root, 'fileAuthor')
-    file_author.text = c.file_author['name']
-
-    # right now ChemKED just uses an integer file version
-    file_version = etree.SubElement(root, 'fileVersion')
-    major_version = etree.SubElement(file_version, 'major')
-    major_version.text = str(c.file_version)
-    minor_version = etree.SubElement(file_version, 'minor')
-    minor_version.text = '0'
-
-    respecth_version = etree.SubElement(root, 'ReSpecThVersion')
-    major_version = etree.SubElement(respecth_version, 'major')
-    major_version.text = '1'
-    minor_version = etree.SubElement(respecth_version, 'minor')
-    minor_version.text = '0'
-
-    # Only ignition delay currently supported
-    exp = etree.SubElement(root, 'experimentType')
-    if c.experiment_type == 'ignition delay':
-        exp.text = 'Ignition delay measurement'
-    else:
-        raise NotImplementedError('Only ignition delay type supported for conversion.')
-
-    reference = etree.SubElement(root, 'bibliographyLink')
-    citation = ''
-    for author in c.reference.authors:
-        citation += author['name'] + ', '
-    citation += (c.reference.journal + ' (' + str(c.reference.year) + ') ' +
-                 str(c.reference.volume) + ':' + c.reference.pages + '. ' + c.reference.detail
-                 )
-    reference.set('preferredKey', citation)
-    reference.set('doi', c.reference.doi)
-
-    apparatus = etree.SubElement(root, 'apparatus')
-    kind = etree.SubElement(apparatus, 'kind')
-    kind.text = c.apparatus.kind
-
-    common_properties = etree.SubElement(root, 'commonProperties')
-    # ChemKED objects have no common properties once loaded. Check for properties
-    # among datapoints that tend to be common
-    common = []
-    composition = c.datapoints[0].composition
-    # Composition type *has* to be the same
-    composition_type = c.datapoints[0].composition_type
-    if all([composition == dp.composition for dp in c.datapoints]):
-        # initial composition is common
-        common.append('composition')
-        prop = etree.SubElement(common_properties, 'property')
-        prop.set('name', 'initial composition')
-
-        for species in composition:
-            component = etree.SubElement(prop, 'component')
-            species_link = etree.SubElement(component, 'speciesLink')
-            species_link.set('preferredKey', species['species-name'])
-            if species.get('InChI'):
-                species_link.set('InChI', species['InChI'])
-
-            amount = etree.SubElement(component, 'amount')
-            amount.set('units', composition_type)
-            amount.text = str(species['amount'].magnitude)
-
-    # If multiple datapoints present, then find any common properties. If only
-    # one datapoint, then composition should be the only "common" property.
-    if len(c.datapoints) > 1:
-        for prop_name in datagroup_properties:
-            attribute = prop_name.replace(' ', '_')
-            quantity = getattr(c.datapoints[0], attribute)
-            if (quantity != None and
-                all([quantity == getattr(dp, attribute) for dp in c.datapoints])
-                ):
-                common.append(prop_name)
-                prop = etree.SubElement(common_properties, 'property')
-                prop.set('description', '')
-                prop.set('name', prop_name)
-                prop.set('units', str(quantity.units))
-
-                value = etree.SubElement(prop, 'value')
-                value.text = str(quantity.magnitude)
-
-    # Ignition delay can't be common, unless only a single datapoint.
-
-    datagroup = etree.SubElement(root, 'dataGroup')
-    datagroup.set('id', 'dg1')
-    datagroup_link = etree.SubElement(datagroup, 'dataGroupLink')
-    datagroup_link.set('dataGroupID', '')
-    datagroup_link.set('dataPointID', '')
-
-    property_idx = {}
-    labels = {'temperature': 'T', 'pressure': 'P',
-              'ignition delay': 'tau', 'pressure rise': 'dP/dt',
-              }
-
-    for prop_name in datagroup_properties:
-        attribute = prop_name.replace(' ', '_')
-        if (prop_name not in common and
-            any([getattr(dp, attribute, None) for dp in c.datapoints])
-            ):
-            prop = etree.SubElement(datagroup, 'property')
-            prop.set('description', '')
-            prop.set('name', prop_name)
-            prop.set('units', str(getattr(c.datapoints[0], attribute).units))
-            idx = 'x{}'.format(len(property_idx) + 1)
-            property_idx[idx] = prop_name
-            prop.set('id', idx)
-            prop.set('label', labels[prop_name])
-
-    if 'composition' not in common:
-        for species in c.datapoints[0].composition:
-            prop = etree.SubElement(datagroup, 'property')
-            prop.set('description', '')
-
-            idx = 'x{}'.format(len(property_idx) + 1)
-            property_idx[idx] = species['species-name']
-            prop.set('id', idx)
-            prop.set('label', '[' + species['species-name'] + ']')
-            prop.set('name', 'composition')
-            prop.set('units', c.datapoints[0].composition_type)
-
-            species_link = etree.SubElement(prop, 'speciesLink')
-            species_link.set('preferredKey', species['species-name'])
-            if species.get('InChI'):
-                species_link.set('InChI', species['InChI'])
-
-    for dp in c.datapoints:
-        datapoint = etree.SubElement(datagroup, 'dataPoint')
-        for idx in property_idx:
-            value = etree.SubElement(datapoint, idx)
-            value.text = str(getattr(dp, property_idx[idx].replace(' ', '_')).magnitude)
-
-    # if RCM and has volume history, need a second dataGroup
-    if (len(c.datapoints) > 1 and
-        any([getattr(dp, 'volume_history', None) for dp in c.datapoints])
-        ):
-        raise NotImplementedError('Error: ReSpecTh files do not support multiple datapoints '
-                                  'with a volume history.'
-                                  )
-        # TODO: what if they share the same history? Does this happen?
-    elif getattr(c.datapoints[0], 'volume_history', None):
-        datagroup = etree.SubElement(root, 'dataGroup')
-        datagroup.set('id', 'dg1')
-        datagroup_link = etree.SubElement(datagroup, 'dataGroupLink')
-        datagroup_link.set('dataGroupID', '')
-        datagroup_link.set('dataPointID', '')
-
-        # Volume history has two properties: time and volume.
-        volume_history = c.datapoints[0].volume_history
-        prop = etree.SubElement(datagroup, 'property')
-        prop.set('description', '')
-        prop.set('name', 'time')
-        prop.set('units', str(volume_history.time.units))
-        time_idx = 'x{}'.format(len(property_idx) + 1)
-        prop.set('id', time_idx)
-        prop.set('label', 't')
-
-        prop = etree.SubElement(datagroup, 'property')
-        prop.set('description', '')
-        prop.set('name', 'volume')
-        prop.set('units', str(volume_history.volume.units))
-        volume_idx = 'x{}'.format(len(property_idx) + 2)
-        prop.set('id', volume_idx)
-        prop.set('label', 'V')
-
-        for time, volume in zip(volume_history.time, volume_history.volume):
-            datapoint = etree.SubElement(datagroup, 'dataPoint')
-            value = etree.SubElement(datapoint, time_idx)
-            value.text = str(time.magnitude)
-            value = etree.SubElement(datapoint, volume_idx)
-            value.text = str(volume.magnitude)
-
-    # In ReSpecTh files all datapoints share ignition type
-    ignition = etree.SubElement(root, 'ignitionType')
-    if c.datapoints[0].ignition_type['target'] == 'pressure':
-        ignition.set('target', 'P')
-    elif c.datapoints[0].ignition_type['target'] == 'temperature':
-        ignition.set('target', 'T')
-    else:
-        # options left are species
-        ignition.set('target', c.datapoints[0].ignition_type['target'])
-    ignition.set('type', c.datapoints[0].ignition_type['type'])
-
-    et = etree.ElementTree(root)
-    if not filename_xml:
-        filename_xml = os.path.splitext(os.path.basename(filename_ck))[0] + '.xml'
-
-    et.write(filename_xml, encoding='utf-8', xml_declaration=True)
-    print('Converted to ' + filename_xml)
-
-
 def main(argv):
     """
     """
+    from pyked.chemked import ChemKED
     parser = ArgumentParser(
         description='Convert between ReSpecTh XML file and ChemKED YAML file '
                     'automatically based on file extension.'
@@ -697,19 +496,24 @@ def main(argv):
     if (os.path.splitext(args.input)[1] == '.xml' and
         os.path.splitext(args.output)[1] == '.yaml'
         ):
-        convert_from_ReSpecTh(args.input, args.output, args.file_author, args.file_author_orcid)
+        ReSpecTh_to_ChemKED(args.input, args.output, args.file_author, args.file_author_orcid)
+
     elif (os.path.splitext(args.input)[1] == '.yaml' and
           os.path.splitext(args.output)[1] == '.xml'
           ):
-        convert_to_ReSpecTh(args.input, args.output)
+        c = ChemKED(yaml_file=args.input)
+        c.convert_to_ReSpecTh(args.output)
+
     elif (os.path.splitext(args.input)[1] == '.xml' and
           os.path.splitext(args.output)[1] == '.xml'
           ):
         raise KeywordError('Cannot convert .xml to .xml')
+
     elif (os.path.splitext(args.input)[1] == '.yaml' and
           os.path.splitext(args.output)[1] == '.yaml'
           ):
         raise KeywordError('Cannot convert .yaml to .yaml')
+
     else:
         raise KeywordError('Input/output args need to be .xml/.yaml')
